+++ src/PowerModels.jl
@@ -43,6 +43,7 @@ include("io/pti.jl")
 include("io/psse.jl")
 
 include("core/data.jl")
+include("core/data_basic.jl")
 include("core/ref.jl")
 include("core/base.jl")
 include("core/types.jl")