+++ CHANGELOG.md
@@ -2,11 +2,13 @@ PowerModels.jl Change Log
 =========================
 
 ### Staged
+- Added support for matrix-based analysis of basic network data (#728)
 - Added support for `relax_integrality` in `run_model`
 - Added `export_pti` to write a PSSE file (#752)
 - Added `parse_files` to create a PM-multinetwork from multiples files
-- Add support for convering matpower ramp rates into per-unit (#561)
+- Added support for convering matpower ramp rates into per-unit (#561)
 - Fixed bug in dual reporting in `constraint_power_balance_ls` (#741)
+- Fixed sign convetion for power injection in `calc_bus_injection`
 
 ### v0.17.3
 - Added a to file variant of `export_matpower`